target dir is checked in
Reference target folder with important files is checked in. This include sqlitejdbc.dll for Windows-11, libsqlitejdbc.so for Linux (Ubuntu 24.04) and common jar file sqlcipher-jdbc.jar. The common jar file is tested and deployed on servlets on Windows-11 and Linux (Ubuntu 24.04). Hence that common jar file can be used as is on those platforms.